requirements.txt
pbr==0.11.1
iso8601
eventlet
netaddr>=0.7.6
pyserial
oslo.config==1.15.0
six>=1.7.0
Babel>=1.3
oauthlib
netifaces
PyYAML
tzlocal
